Use same settings dialog for both screens
