+++ packages/devtools_app/lib/src/screens/inspector/inspector_screen.dart
@@ -8,7 +8,6 @@ import 'dart:collection';
 import 'package:devtools_app_shared/ui.dart';
 import 'package:devtools_app_shared/utils.dart';
 import 'package:flutter/material.dart';
-import 'package:vm_service/vm_service.dart' hide Stack;
 
 import '../../service/service_extension_widgets.dart';
 import '../../service/service_extensions.dart' as extensions;
@@ -16,14 +15,12 @@ import '../../shared/analytics/analytics.dart' as ga;
 import '../../shared/analytics/constants.dart' as gac;
 import '../../shared/common_widgets.dart';
 import '../../shared/console/eval/inspector_tree.dart';
-import '../../shared/editable_list.dart';
 import '../../shared/error_badge_manager.dart';
 import '../../shared/globals.dart';
-import '../../shared/preferences/preferences.dart';
 import '../../shared/primitives/blocking_action_mixin.dart';
-import '../../shared/primitives/simple_items.dart';
 import '../../shared/ui/search.dart';
 import '../inspector_shared/inspector_screen.dart';
+import '../inspector_shared/inspector_settings_dialog.dart';
 import 'inspector_controller.dart';
 import 'inspector_screen_details_tab.dart';
 import 'inspector_tree_controller.dart';
@@ -314,80 +311,6 @@ class InspectorScreenBodyState extends State<InspectorScreenBody>
   }
 }
 
-class FlutterInspectorSettingsDialog extends StatelessWidget {
-  const FlutterInspectorSettingsDialog({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final dialogHeight = scaleByFontFactor(500.0);
-    return DevToolsDialog(
-      title: const DialogTitleText('Flutter Inspector Settings'),
-      content: SizedBox(
-        width: defaultDialogWidth,
-        height: dialogHeight,
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            ...dialogSubHeader(
-              theme,
-              'General',
-            ),
-            CheckboxSetting(
-              notifier: preferences.inspector.hoverEvalModeEnabled
-                  as ValueNotifier<bool?>,
-              title: 'Enable hover inspection',
-              description:
-                  'Hovering over any widget displays its properties and values.',
-              gaItem: gac.inspectorHoverEvalMode,
-            ),
-            CheckboxSetting(
-              notifier: preferences.inspector.inspectorV2Enabled
-                  as ValueNotifier<bool?>,
-              title: 'Enable Inspector V2',
-              description: 'Try out the new Inspector screen.',
-              gaItem: gac.inspectorV2Enabled,
-            ),
-            const SizedBox(height: denseSpacing),
-            const InspectorDefaultDetailsViewOption(),
-            const SizedBox(height: denseSpacing),
-            ...dialogSubHeader(theme, 'Package Directories'),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                Expanded(
-                  child: Text(
-                    'Widgets in these directories will show up in your summary tree.',
-                    style: theme.subtleTextStyle,
-                  ),
-                ),
-                MoreInfoLink(
-                  url: DocLinks.inspectorPackageDirectories.value,
-                  gaScreenName: gac.inspector,
-                  gaSelectedItemDescription:
-                      gac.InspectorDocs.packageDirectoriesDocs.name,
-                ),
-              ],
-            ),
-            Text(
-              '(e.g. /absolute/path/to/myPackage/)',
-              style: theme.subtleTextStyle,
-            ),
-            const SizedBox(height: denseSpacing),
-            const Expanded(
-              child: PubRootDirectorySection(),
-            ),
-          ],
-        ),
-      ),
-      actions: const [
-        DialogCloseButton(),
-      ],
-    );
-  }
-}
-
 class InspectorSummaryTreeControls extends StatelessWidget {
   const InspectorSummaryTreeControls({
     super.key,
@@ -562,97 +485,6 @@ class _ErrorNavigatorButton extends StatelessWidget {
         color: Theme.of(context).colorScheme.onErrorContainer,
         onPressed: onPressed,
       ),
-    );
-  }
-}
-
-class InspectorDefaultDetailsViewOption extends StatelessWidget {
-  const InspectorDefaultDetailsViewOption({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return ValueListenableBuilder(
-      valueListenable: preferences.inspector.defaultDetailsView,
-      builder: (context, selection, _) {
-        final theme = Theme.of(context);
-        return Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Text(
-              'Select the default tab for the inspector.',
-              style: theme.subtleTextStyle,
-            ),
-            const SizedBox(height: denseSpacing),
-            Row(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                Radio<InspectorDetailsViewType>(
-                  value: InspectorDetailsViewType.layoutExplorer,
-                  materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                  groupValue: selection,
-                  onChanged: _onChanged,
-                ),
-                Text(InspectorDetailsViewType.layoutExplorer.key),
-                const SizedBox(width: denseSpacing),
-                Radio<InspectorDetailsViewType>(
-                  materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                  value: InspectorDetailsViewType.widgetDetailsTree,
-                  groupValue: selection,
-                  onChanged: _onChanged,
-                ),
-                Text(InspectorDetailsViewType.widgetDetailsTree.key),
-              ],
-            ),
-          ],
-        );
-      },
-    );
-  }
-
-  void _onChanged(InspectorDetailsViewType? value) {
-    if (value != null) {
-      preferences.inspector.setDefaultInspectorDetailsView(value);
-      final item = value.name == InspectorDetailsViewType.layoutExplorer.name
-          ? gac.defaultDetailsViewToLayoutExplorer
-          : gac.defaultDetailsViewToWidgetDetails;
-      ga.select(
-        gac.inspector,
-        item,
-      );
-    }
-  }
-}
-
-class PubRootDirectorySection extends StatelessWidget {
-  const PubRootDirectorySection({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return ValueListenableBuilder<IsolateRef?>(
-      valueListenable:
-          serviceConnection.serviceManager.isolateManager.mainIsolate,
-      builder: (_, __, ___) {
-        return SizedBox(
-          height: 200.0,
-          child: EditableList(
-            gaScreen: gac.inspector,
-            gaRefreshSelection: gac.refreshPubRoots,
-            entries: preferences.inspector.pubRootDirectories,
-            textFieldLabel: 'Enter a new package directory',
-            isRefreshing: preferences.inspector.isRefreshingPubRootDirectories,
-            onEntryAdded: (p0) => unawaited(
-              preferences.inspector.addPubRootDirectories(
-                [p0],
-                shouldCache: true,
-              ),
-            ),
-            onEntryRemoved: (p0) =>
-                unawaited(preferences.inspector.removePubRootDirectories([p0])),
-            onRefreshTriggered: () =>
-                unawaited(preferences.inspector.loadPubRootDirectories()),
-          ),
-        );
-      },
     );
   }
 }
