+++ packages/devtools_app/lib/src/screens/inspector_v2/inspector_controls.dart
@@ -12,6 +12,7 @@ import '../../service/service_extensions.dart' as extensions;
 import '../../shared/analytics/constants.dart' as gac;
 import '../../shared/common_widgets.dart';
 import '../../shared/globals.dart';
+import '../inspector_shared/inspector_settings_dialog.dart';
 import 'inspector_controller.dart';
 import 'inspector_screen.dart';
